test: ensure compile output works correctly
