+++ Cargo.toml
@@ -181,7 +181,8 @@ vergen = { version = "8.3.1", features = [
 # Testing
 dhat = "0.3.3"
 divan = "0.1.14"
-insta = { version = "1.39", features = ["glob"] }
+insta = { version = "1.43", features = ["glob"] }
+insta-cmd = "0.6.0"
 
 
 # Our Own Crates
