+++ tests/e2e/e2e.rs
@@ -1,0 +1,2 @@
+mod cli;
+mod lsp;
